usb: dwc3: Fix warnings on 64-bit builds
Change aritmentics to use 64bit types to be compatible with 64bit
builds.

Compilation warnings:
drivers/usb/dwc3/ep0.c: In function ‘dwc3_ep0_start_trans’:
drivers/usb/dwc3/ep0.c:85:19: warning: cast from pointer to integer of
different size [-Wpointer-to-int-cast]
  dwc3_flush_cache((int)trb, sizeof(*trb));
                   ^
drivers/usb/dwc3/gadget.c: In function ‘dwc3_prepare_one_trb’:
drivers/usb/dwc3/gadget.c:775:19: warning: cast from pointer to integer
of different size [-Wpointer-to-int-cast]
  dwc3_flush_cache((int)trb, sizeof(*trb));
                   ^
drivers/usb/dwc3/core.c: In function ‘dwc3_setup_scratch_buffers’:
drivers/usb/dwc3/core.c:284:19: warning: cast to pointer from integer of
different size [-Wint-to-pointer-cast]
  dma_unmap_single((void *)dwc->scratch_addr, dwc->nr_scratch *
                   ^
drivers/usb/dwc3/core.c: In function ‘dwc3_free_scratch_buffers’:
drivers/usb/dwc3/core.c:299:19: warning: cast to pointer from integer of
different size [-Wint-to-pointer-cast]
  dma_unmap_single((void *)dwc->scratch_addr, dwc->nr_scratch *
                   ^
drivers/usb/dwc3/core.c: In function ‘dwc3_uboot_init’:
drivers/usb/dwc3/core.c:632:14: warning: cast to pointer from integer of
different size [-Wint-to-pointer-cast]
  dwc->regs = (int *)(dwc3_dev->base + DWC3_GLOBALS_REGS_START);
              ^
drivers/usb/dwc3/ep0.c: In function ‘dwc3_ep0_complete_data’:
drivers/usb/dwc3/ep0.c:793:19: warning: cast from pointer to integer of
different size [-Wpointer-to-int-cast]
  dwc3_flush_cache((int)trb, sizeof(*trb));
                   ^
drivers/usb/dwc3/ep0.c:824:21: warning: cast from pointer to integer of
different size [-Wpointer-to-int-cast]
    dwc3_flush_cache((int)trb, sizeof(*trb));
                     ^
drivers/usb/dwc3/ep0.c:834:20: warning: cast from pointer to integer of
different size [-Wpointer-to-int-cast]
   dwc3_flush_cache((int)dwc->ep0_bounce, DWC3_EP0_BOUNCE_SIZE);
                    ^
  CC      lib/display_options.o
drivers/usb/dwc3/gadget.c: In function ‘dwc3_cleanup_done_reqs’:
drivers/usb/dwc3/gadget.c:1772:19: warning: cast from pointer to integer
of different size [-Wpointer-to-int-cast]
  dwc3_flush_cache((int)trb, sizeof(*trb));
                   ^
drivers/usb/dwc3/gadget.c: In function
‘dwc3_gadget_uboot_handle_interrupt’:
drivers/usb/dwc3/gadget.c:2673:21: warning: cast from pointer to integer
of different size [-Wpointer-to-int-cast]
    dwc3_flush_cache((int)evt->buf, evt->length);
                     ^

+++ drivers/usb/dwc3/io.h
@@ -23,7 +23,7 @@
 #define	CACHELINE_SIZE		CONFIG_SYS_CACHELINE_SIZE
 static inline u32 dwc3_readl(void __iomem *base, u32 offset)
 {
-	u32 offs = offset - DWC3_GLOBALS_REGS_START;
+	unsigned long offs = offset - DWC3_GLOBALS_REGS_START;
 	u32 value;
 
 	/*
@@ -31,7 +31,7 @@ static inline u32 dwc3_readl(void __iomem *base, u32 offset)
 	 * space, see dwc3_probe in core.c.
 	 * However, the offsets are given starting from xHCI address space.
 	 */
-	offs += base;
+	offs += (unsigned long)base;
 	value = readl(offs);
 
 	return value;
@@ -39,14 +39,14 @@ static inline u32 dwc3_readl(void __iomem *base, u32 offset)
 
 static inline void dwc3_writel(void __iomem *base, u32 offset, u32 value)
 {
-	u32 offs = offset - DWC3_GLOBALS_REGS_START;
+	unsigned long offs = offset - DWC3_GLOBALS_REGS_START;
 
 	/*
 	 * We requested the mem region starting from the Globals address
 	 * space, see dwc3_probe in core.c.
 	 * However, the offsets are given starting from xHCI address space.
 	 */
-	offs += base;
+	offs += (unsigned long)base;
 	writel(value, offs);
 }
